rename data.rs to csv_reader.rs
